requirements.txt
absl-py==1.2.0
asgiref==3.8.1
asttokens==2.0.8
astunparse==1.6.3
backcall==0.2.0
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
distlib==0.3.7
Django==5.2
django-filter==25.1
django-material-admin==1.8.6
djangorestframework==3.16.0
entrypoints==0.4
executing==1.1.1
filelock==3.12.3
Flask==2.2.2
flatbuffers==1.12
fonttools==4.37.2
gast==0.4.0
google-auth==2.10.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
gunicorn==23.0.0
h5py==3.7.0
idna==3.3
install==1.3.5
ipykernel==6.16.0
ipython==8.5.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
jupyter-core==4.11.1
jupyter_client==7.4.2
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mysql-connector-python==8.3.0
nest-asyncio==1.5.6
numpy==1.23.1
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==3.10.0
prompt-toolkit==3.0.31
protobuf==3.19.4
psutil==5.9.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
stack-data==0.5.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
torch==1.12.1
tornado==6.2
traitlets==5.4.0
typing_extensions==4.7.1
tzdata==2022.7
urllib3==1.26.11
virtualenv==20.24.4
wcwidth==0.2.5
Werkzeug==2.2.2
wrapt==1.14.1
